Reorganize original helpers
