+++ source/components/switch.scss
@@ -2,7 +2,7 @@
 // -----------------------------------------------------------------------------
 
 @use "../settings" as *;
-@use "../helpers/focus-visible" as *;
+@use "../helpers/accessibility/focus-visible" as *;
 
 .switch {
   width: 2.5rem;
