+++ source/helpers/accessibility/trap-focus.js
