+++ source/reset.scss
@@ -2,7 +2,7 @@
 // -----------------------------------------------------------------------------
 
 @use "settings" as *;
-@use "helpers/focus-visible" as *;
+@use "helpers/accessibility/focus-visible" as *;
 
 // Universal
